Layout improvements; SearchBox for filtering owned games by name; Other improvements

+++ GamePlatformsClient/Model/SteamResponseData/GetSchemaForGameData.cs
@@ -39,7 +39,7 @@ namespace GamePlatformsClient.Model.SteamResponseData
         public class Stat
         {
             [JsonPropertyName("name")] public string Name { get; set; }
-            [JsonPropertyName("defaultvalue")] public int Defaultvalue { get; set; }
+            [JsonPropertyName("defaultvalue")] public Int64 Defaultvalue { get; set; }
             [JsonPropertyName("displayName")] public string DisplayName { get; set; }
         }
